% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{add_upr}
\alias{add_upr}
\title{Add column containing upper limit for censored poisson}
\usage{
add_upr(
  dat,
  prop_removed_col,
  n_catch_col,
  n_hooks_col,
  pstar_col = "pstar",
  pstar = NULL
)
}
\arguments{
\item{dat}{A data frame from \code{\link[=prep_stitch_dat]{prep_stitch_dat()}}.}

\item{prop_removed_col}{Name of the column containing the proportion of
baits removed in each fishing event from \emph{any} species. I.e., the proportion
of hooks returning without bait for any reason.}

\item{n_catch_col}{Name of the column containing the observed catch counts on
each fishing event of the target species.}

\item{n_hooks_col}{Name of the column containing the number of hooks retrieved
on each fishing event.}

\item{pstar_col}{Name of the column containing a single value between
\verb{0 <= pstar <= 1} specifying the breakdown point of observed catch counts as
a result of hook competition. See \code{\link[=get_pstar]{get_pstar()}}, \verb{[sdmTMB::censored_poisson()]}}

\item{pstar}{Optional. If \code{pstar_col} is not specified, pstar can be provided
as a single value between 0 <= pstar <= 1` default = NULL.}
}
\value{
\code{dat} with a new column \code{upr} containing  numeric vector of upper
bound catch counts of the target species to improve convergence of the censored
method. See the documentation in \verb{[sdmTMB::get_censored_upper()]}
}
\description{
Add column containing upper limit for censored poisson
}
