% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pstar.R
\name{central_diff}
\alias{central_diff}
\title{Calculate the Central Difference Approximation of the First Derivative}
\usage{
central_diff(gam_object, newdata, x, h)
}
\arguments{
\item{gam_object}{A fitted GAM model.}

\item{newdata}{A data frame containing the values of predictor variables
used to calculate the derivative.}

\item{x}{Name of the predictor variable with respect to which you want to
calculate the derivative.}

\item{h}{Step size for the central difference approximation (default = 0.005)}
}
\value{
A numeric vector representing the central difference approximation of
the first derivative.
}
\description{
This function calculates the central difference approximation of the first
derivative of a generalized additive model (GAM) with respect to a specified
predictor variable.
}
