% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_name}
\alias{clean_name}
\title{Clean species names for file names}
\usage{
clean_name(x)
}
\arguments{
\item{x}{species name(s) with slashes and spaces}
}
\value{
Species name(s) that work well for file names
}
\description{
Clean species names for file names
}
\details{
Exported because it turns out this is useful in scripts all over the place.
}
\examples{
clean_name("pacific cod")
}
