% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{fit_survey_maps}
\alias{fit_survey_maps}
\title{Fit geostatistical survey models and make maps}
\usage{
fit_survey_maps(
  dat,
  species = "pacific cod",
  include_depth = TRUE,
  model = c("inla", "glmmfields", "sdmTMB"),
  surveys = c("SYN QCS", "SYN HS", "SYN WCHG", "SYN WCVI"),
  years = c(2017, 2018),
  ...
)
}
\arguments{
\item{dat}{Data from \code{\link[gfplot:get_survey_sets]{gfplot::get_survey_sets()}}.}

\item{species}{The species name.}

\item{include_depth}{Logical for whether or not to include depth and depth
squared as predictors.}

\item{model}{Whether to fit an INLA or glmmfields model.}

\item{surveys}{Character vector describing the surveys to fit. Can include
multiple surveys if desired.}

\item{years}{The years of survey data to use.}

\item{...}{Any other arguments to pass to \code{\link[gfplot:fit_survey_sets]{gfplot::fit_survey_sets()}}.}
}
\description{
\code{fit_survey_maps()} is a helper function that calls
\code{\link[gfplot:fit_survey_sets]{gfplot::fit_survey_sets()}} appropriately for the various surveys.
}
