% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plotpages_Rmd.R
\name{generate_plotpages_Rmd}
\alias{generate_plotpages_Rmd}
\title{Generate plot-pages.Rmd for a species}
\usage{
generate_plotpages_Rmd(x, spp)
}
\arguments{
\item{x}{A species common name}

\item{spp}{Output from \code{\link[=get_spp_names]{get_spp_names()}} that has been cleaned by other
functions.}
}
\value{
Rmd code in a character vector.
}
\description{
Generate plot-pages.Rmd for a species
}
