% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Get all data (includes some IPHC data that will get over-written
in Survey biomass indices section of make-pages.R)}
\usage{
get_data(
  type = c("A", "B"),
  path = ".",
  compress = "gzip",
  force = FALSE,
  sleep = 2
)
}
\arguments{
\item{type}{A or B}

\item{path}{Path}

\item{compress}{Compression passed to \code{saveRDS()}. Defaults to \code{"gzip"}.}

\item{force}{Should data be downloaded even if already cached data exists?}

\item{sleep}{System sleep in seconds between each species
to be nice to the server.}
}
\description{
Get all data (includes some IPHC data that will get over-written
in Survey biomass indices section of make-pages.R)
}
