% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_iphc}
\alias{get_data_iphc}
\title{Get the IPHC data for all years, should get merged into get_data at some point
but before gfsynopsis meeting we don't want to re-get all the non-IPHC data}
\usage{
get_data_iphc(type = c("A", "B"), path = ".", compress = FALSE, force = FALSE)
}
\arguments{
\item{type}{A or B}

\item{path}{Path}

\item{compress}{TRUE or FALSE to compress .rds}

\item{force}{Should data be downloaded even if already cached data exists?}
}
\description{
Get the IPHC data for all years, should get merged into get_data at some point
but before gfsynopsis meeting we don't want to re-get all the non-IPHC data
}
