% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_iphc_hook_with_bait}
\alias{get_data_iphc_hook_with_bait}
\title{Get the IPHC data for all years for hooks with bait (for hook competition
calculations.}
\usage{
get_data_iphc_hook_with_bait(path = ".", compress = FALSE, force = FALSE)
}
\arguments{
\item{path}{Path}

\item{compress}{TRUE or FALSE to compress .rds}

\item{force}{Should data be downloaded even if already cached data exists?}
}
\description{
Get the IPHC data for all years for hooks with bait (for hook competition
calculations.
}
