% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{get_family_object}
\alias{get_family_object}
\title{Retrieve an R Family Object from a String}
\usage{
get_family_object(fit_fam)
}
\arguments{
\item{fit_fam}{A character string specifying the family to use.
Must be one of the following:
\itemize{
\item \code{"delta-gamma"}: Delta-Gamma distribution (\code{sdmTMB::delta_gamma})
\item \code{"delta-lognormal"}: Delta-Lognormal distribution (\code{sdmTMB::delta_lognormal})
\item \code{"delta-gengamma"}: Delta-Generalized Gamma distribution (\code{sdmTMB::delta_gengamma})
\item \code{"delta-gamma-poisson-link"}: Delta-Gamma with Poisson link (\code{sdmTMB::delta_gamma})
\item \code{"delta-lognormal-poisson-link"}: Delta-Lognormal with Poisson link (\code{sdmTMB::delta_lognormal})
\item \code{"delta-gengamma-poisson-link"}: Delta-Generalized Gamma with Poisson link (\code{sdmTMB::delta_gengamma})
\item \code{"tweedie"}: Tweedie distribution (\code{sdmTMB::tweedie})
\item \code{"nb2"}: Negative binomial (\code{sdmTMB::nbinom2})
}}
}
\value{
An R family object used for model fitting.
}
\description{
This function takes a character string representing a distribution family name
and returns the corresponding R family object, primarily for use in statistical
modeling with \code{sdmTMB} or \code{glm}. The default link function for all distributions
is \code{"log"}, unless otherwise specified.
}
\examples{
get_family_object("lognormal")
}
