% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pstar.R
\name{get_fderiv}
\alias{get_fderiv}
\title{Calculate First Derivative and Confidence Intervals for GAM Smooth Terms}
\usage{
get_fderiv(
  gam_object,
  newdata,
  terms,
  term_columns,
  x,
  h = 0.005,
  nsim,
  level = 0.95
)
}
\arguments{
\item{gam_object}{A fitted GAM model.}

\item{newdata}{A data frame containing the values of predictor variables
to calculate the derivative.}

\item{terms}{A character vector specifying the smooth terms for which
derivatives are calculated.}

\item{term_columns}{A character vector specifying the columns in 'newdata' that
correspond to the terms specified in 'terms'.}

\item{x}{Name of the predictor variable with respect to which you want to
calculate the derivatives.}

\item{h}{Step size for the central difference approximation (default = 0.005)}

\item{nsim}{The number of simulations used to estimate confidence intervals.}

\item{level}{Confidence level for the intervals (default is 0.95).}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{smooth}: Name of the smooth term.
\item \code{newdata}: Values of the corresponding term in 'newdata'.
\item \code{derivative}: Point estimate of the derivative.
\item \code{se}: Standard error of the estimate.
\item \code{crit}: Critical value for the confidence interval.
\item \code{lower}: Lower bound of the confidence interval.
\item \code{upper}: Upper bound of the confidence interval.
}
}
\description{
This function calculates the first derivatives and confidence intervals for
specified smooth terms in a generalized additive model (GAM) using central
difference approximations. It provides point estimates and confidence
intervals for the derivatives of the specified terms.
}
