% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{get_inclusion_table}
\alias{get_inclusion_table}
\title{Get proportion of positive sets in each SYN or HBLL region}
\usage{
get_inclusion_table(survey_dat = NULL, survey_type)
}
\arguments{
\item{survey_dat}{A dataframe from \verb{[gfplot::get_survey_sets()]}}

\item{survey_type}{A string matching one of: "synoptic", "hbll_outside", "hbll_inside"}
}
\value{
A \code{flextable} object in a browser.
}
\description{
A lookup table to see the proportion of positive sets across species and regions.
Renders a \code{flextable} object in a browser with rows highlighted based on
the proportion of positive sets ("red" less than or equal to 0.01, "orange"
between 0.01 and 0.03, "yellow" between 0.03 and 0.05).
}
