% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_iphc_hook_data}
\alias{get_iphc_hook_data}
\title{Get IPHC hook counts for all survey years}
\usage{
get_iphc_hook_data(path = ".", species = "pacific halibut")
}
\arguments{
\item{path}{Path where the 'iphc-hook-counts.rds' file is saved to}

\item{species}{Species common name to query, defaults to 'pacific halibut'
because to date this species does not affect the output.}
}
\value{
An RDS object containing a dataframe with the following columns:
\itemize{
\item year: The year of the data.
\item station: The station identifier.
\item lat: The latitude coordinates.
\item lon: The longitude coordinates.
\item baited_hooks: The number of baited hooks.
\item setID: The set identifier (used by the IPHC, may not match GFBio output)
\item obsHooksPerSet: The observed hooks per set.
\item effSkateIPHC: The effective skate IPHC (International Pacific Halibut Commission).
\item iphcUsabilityCode: The IPHC usability code.
\item iphcUsabilityDesc: The description of the IPHC usability code.
}
}
\description{
Get IPHC hook counts for all survey years
}
