% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-iphc.R
\name{get_iphc_pos_sets}
\alias{get_iphc_pos_sets}
\title{Get table of positive sets for IPHC FISS data}
\usage{
get_iphc_pos_sets(clean_iphc_dat)
}
\arguments{
\item{survey_dat}{A dataframe from \code{\link[=prep_iphc_stitch_dat]{prep_iphc_stitch_dat()}}}
}
\value{
A dataframe
}
\description{
Get table of positive sets for IPHC FISS data
}
