% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-iphc.R
\name{get_iphc_stitched_index}
\alias{get_iphc_stitched_index}
\title{Get index using \code{sdmTMB} for IPHC surveys}
\usage{
get_iphc_stitched_index(
  survey_dat,
  species,
  model_type = "st-rw",
  form = NULL,
  family = sdmTMB::tweedie(),
  time = "year",
  spatial = "on",
  spatiotemporal = "rw",
  time_varying = NULL,
  time_varying_type = NULL,
  data = survey_dat,
  mesh = NULL,
  cutoff = 20,
  offset = "log_eff_skate",
  extra_time = missing_years,
  priors = sdmTMB::sdmTMBpriors(),
  silent = TRUE,
  ctrl = sdmTMB::sdmTMBcontrol(),
  gradient_thresh = 0.001,
  cache = NULL,
  check_cache = FALSE,
  grid
)
}
\arguments{
\item{survey_dat}{A data frame from \code{\link[=prep_iphc_stitch_dat]{prep_iphc_stitch_dat()}}.}

\item{species}{A string specifying the \code{species_common_name}.}

\item{model_type}{An suffix to the filename indicating model type (default = "st-rw").}

\item{family}{The family and link for \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{mesh}{Optional mesh object created using \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{cutoff}{If \code{mesh = NULL}, mesh cutoff for \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{offset}{A string naming the offset column in \code{dat} used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}} (default = 'log_eff_skate').}

\item{silent}{A boolean. Silent or include optimization details.}

\item{ctrl}{Optimization control options via \code{\link[sdmTMB:sdmTMBcontrol]{sdmTMB::sdmTMBcontrol()}}.}

\item{gradient_thresh}{Threshold used in \code{\link[sdmTMB:sanity]{sdmTMB::sanity()}} (default = 0.001).}

\item{cache}{A string specifying file path to cache directory.}

\item{grid}{A data frame containing the locations of IPHC stations used to make predictions and generate the index.}
}
\value{
Either a string or dataframe:
\itemize{
\item \verb{insufficient data to stitch regions} if the number of positive sets is too low to stitch
\item \verb{Failed sanity check} if the model failed to converge
\item A dataframe containing the stitched index formatted to use with \code{\link[gfplot:plot_survey_index]{gfplot::plot_survey_index()}}
}
}
\description{
Get index using \code{sdmTMB} for IPHC surveys
}
