% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ll_bait_counts}
\alias{get_ll_bait_counts}
\title{Download longline baited hook counts}
\usage{
get_ll_bait_counts(path = ".", species = 442, ssid = c(22, 36, 39, 40))
}
\arguments{
\item{path}{Path where the 'bait-counts.rds' file is saved to}

\item{species}{Species code or common name to query}

\item{ssid}{Survey series id defaulting to the longline surveys}
}
\value{
RDS object with a stored dataframe ssid, year, fishing_event_id, and baited hook count
}
\description{
Get the hook data for the HBLL outside and inside surveys from GFBIO.
}
