% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pstar.R
\name{get_pstar}
\alias{get_pstar}
\title{Get pstar - Estimate Breakdown Point of Catch Counts}
\usage{
get_pstar(
  survey_dat,
  gam_formula,
  survey_type,
  prop_removed_min = NULL,
  h = 0.005,
  pstar_cache = NULL,
  save_out = TRUE
)
}
\arguments{
\item{survey_dat}{A dataframe from \code{\link[=prep_stitch_dat]{prep_stitch_dat()}} or
\code{\link[=prep_iphc_stitch_dat]{prep_iphc_stitch_dat()}}.}

\item{gam_formula}{A string or formula specifying the generalised additive
used to estimate pstar. (See Watson et al. 2023).
\itemize{
\item Suggestion for IPHC \code{catch ~ -1 + s(prop_removed) + fyear + s(fstation, bs = "re") + offset(log_eff_skate)}
\item Suggestion for HBLL \code{catch ~ -1 + s(prop_removed) + fyear + offset(log(hook_count))}.
}}

\item{survey_type}{A string specifying the survey type: "iphc",
"hbll_outside", or "hbll_inside".}

\item{prop_removed_min}{Optional. Minimum value of proportion of baits removed
for generating predictions. If not provided, uses the minimum value found in
\code{survey_dat}.}

\item{h}{Step size for the central difference approximation (default = 0.005).}

\item{pstar_cache}{Path to a folder for caching pstar objects.}

\item{save_out}{Whether to save the pstar object to \code{pstar_cache}
(default = TRUE).

#' @references
Watson, J., Edwards, A.M., and Auger-Méthe M. 2023. A statistical censoring
approach accounts for hook competition in abundance indices from longline
surveys. Can. J. Fish. Aquat. Sci. 80(3): 468-486. \doi{10.1139/cjfas-2022-0159}}
}
\value{
A list of objects containing
\itemize{
\item \code{gam_fit}: The fitted GAM object.
\item \code{pred_df}: A data frame with predicted values from \code{gam_fit}
\item \code{f1}: A data frame containing the numerical approximation of the first derivative
\item \code{pstar_df}: A data frame with the estimated pstar value.
\item \code{species}: A character vector containing the species name.
}
}
\description{
Fits a Generalized Additive Model (GAM) and calculates the breakdown point
(pstar) of observed catch counts due to hook competition.
}
