% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{get_stitched_index}
\alias{get_stitched_index}
\title{Get stitched index across survey regions in synoptic trawl and HBLL surveys}
\usage{
get_stitched_index(
  survey_dat,
  species = "arrowtooth flounder",
  survey_type = "synoptic",
  form = NULL,
  family = "tweedie",
  time = "year",
  spatial = "on",
  spatiotemporal = "rw",
  use_extra_time = NULL,
  time_varying = NULL,
  time_varying_type = NULL,
  mesh = NULL,
  cutoff = 20,
  offset = "offset",
  priors = sdmTMB::sdmTMBpriors(),
  silent = TRUE,
  ctrl = sdmTMB::sdmTMBcontrol(nlminb_loops = 1L, newton_loops = 1L,
    collapse_spatial_variance = TRUE, collapse_threshold = 0.01),
  gradient_thresh = 0.01,
  cache = NULL,
  check_cache = FALSE,
  cache_predictions = FALSE,
  cache_fits = FALSE,
  index_grid = NULL,
  shapefile = NULL
)
}
\arguments{
\item{survey_dat}{A data frame from \code{\link[=prep_stitch_dat]{prep_stitch_dat()}}.}

\item{species}{A string specifying the \code{species_common_name}.}

\item{survey_type}{A string matching one of: "synoptic" (the default), "hbll_outside", "hbll_inside".}

\item{form}{Optional string specifying model formula.
'catch ~ 1' (the default, unless \code{family = poisson()} or \code{family = sdmTMB::censored_possion()}
then "catch ~ 1 + (1|obs_id)").}

\item{family}{A character string specifying the distribution family.
Passed to \code{get_family_object()}.}

\item{time}{An optional time column name (as character), used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
(default = 'year')}

\item{spatial}{Estimate spatial random fields? See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
(Default is 'rw').}

\item{spatiotemporal}{Estimate the spatiotemporal random fields, see \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
(Default is 'rw').}

\item{time_varying}{An optional one-sided formula describing covariates that
should be modelled as a time-varying process. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
Default is \code{NULL} if \code{model_type = st-rw}, and \code{~1} if \code{model_type = 'st-rw_tv-rw'}.}

\item{time_varying_type}{Type of time-varying process to apply to
‘time_varying’ formula. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}. #'    Default is \code{NULL}
if \code{model_type = st-rw}, and \code{~1} if \code{model_type = 'st-rw_tv-rw'}.}

\item{mesh}{Optional mesh object created using \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{cutoff}{If \code{mesh = NULL}, mesh cutoff for \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{offset}{A string naming the offset column in \code{dat} used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{priors}{Optional penalties/priors used in \code{\link[sdmTMB:sdmTMBpriors]{sdmTMB::sdmTMBpriors()}}.}

\item{silent}{A boolean. Silent or include optimization details.}

\item{ctrl}{Optimization control options via \code{\link[sdmTMB:sdmTMBcontrol]{sdmTMB::sdmTMBcontrol()}}.}

\item{gradient_thresh}{Threshold used in \code{\link[sdmTMB:sanity]{sdmTMB::sanity()}} (default = 0.001).}

\item{cache}{A string specifying file path to cache directory.}

\item{check_cache}{Check whether index file already exists? Default = \code{FALSE}.}

\item{cache_predictions}{Cache model predictions? Can be large.}

\item{cache_fits}{Cache model fits? Can be large.}

\item{index_grid}{A data frame containing the spatial grid over which predictions are to be made.
If \code{index_grid} = NULL (the default). Grid should contain cell area.}

\item{shapefile}{An \code{sf} polygon object used to filter the prediction/index grid.
Used for the spatially filtered synopsis report. Default = \code{NULL}}

\item{model_type}{A string matching one of: "st-rw" (the default), "st-rw_tv-rw", or "custom".}
}
\value{
Either a string or dataframe:
\itemize{
\item \verb{insufficient data to stitch regions} if the number of positive sets is too low to stitch
\item \verb{Failed sanity check} if the model failed to converge
\item A dataframe containing the stitched index formatted to use with \code{\link[gfplot:plot_survey_index]{gfplot::plot_survey_index()}}
}
}
\description{
Get stitched index across survey regions in synoptic trawl and HBLL surveys
}
\details{
Allowed values for \code{family}:
\itemize{
\item Standard families: \code{"lognormal"}, \code{"gamma"}, \code{"gengamma"}, \code{"tweedie"}, \code{"nb2"}.
\item Delta models: \code{"delta-gamma"}, \code{"delta-lognormal"}, \code{"delta-gengamma"}.
\item Delta models with Poisson link: \code{"delta-gamma-poisson-link"}, \code{"delta-lognormal-poisson-link"}, \code{"delta-gengamma-poisson-link"}.
}

The selected \code{family} is passed to \code{get_family_object()}, which maps the input string
to the appropriate \code{sdmTMB} family object.
}
