% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfsynopsis-package.R
\docType{package}
\name{gfsynopsis-package}
\alias{gfsynopsis}
\alias{gfsynopsis-package}
\title{gfsynopsis: Data Synopsis Reports for PBS Groundfish}
\description{
Facilitates the creation of reproducible data synopsis report for DFO PBS groundfish stocks in British Columbia, Canada.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seananderson/gfsynopsis}
  \item Report bugs at \url{https://github.com/seananderson/gfsynopsis/issues}
}

}
\author{
\strong{Maintainer}: Sean Anderson \email{sean.anderson@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Elise Keppel
}

}
\keyword{internal}
