% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfsynopsis.R
\docType{package}
\name{gfsynopsis}
\alias{gfsynopsis}
\title{gfsynopsis package}
\description{
gfsynopsis package
}
\details{
See the README on
\href{https://github.com/seananderson/gfsynopsis#readme}{GitHub}
}
