% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot2-like colour scale in HCL space}
\usage{
gg_color_hue(n, hue_min = 8, hue_max = 290, l = 52, c = 100)
}
\arguments{
\item{n}{Number of colours to return.}

\item{hue_min}{Minimum hue value in the range \verb{(0,360)}}

\item{hue_max}{Maximum hue value in the range \verb{(0,360)}}

\item{l}{Luminance in the range \verb{(0,100)}}

\item{c}{Chroma of the colour.}
}
\description{
ggplot2-like colour scale in HCL space
}
\details{
See the \code{\link[grDevices:hcl]{grDevices::hcl()}} function for details.
}
\examples{
gg_color_hue(10)
plot(1:6, col = gg_color_hue(6), pch = 20, cex = 3)
}
