% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_refs_spp.R
\name{join_refs_spp}
\alias{join_refs_spp}
\title{Join reference data onto species data frame and clean up}
\usage{
join_refs_spp(spp, french = FALSE)
}
\arguments{
\item{spp}{Output from \code{\link[=get_spp_names]{get_spp_names()}}}
}
\value{
A modified data frame
}
\description{
Join reference data onto species data frame and clean up
}
\examples{
get_spp_names() |> join_refs_spp()
}
