% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_itis_spp.R
\name{join_worms_spp}
\alias{join_worms_spp}
\title{Join WoRMS data to species data frame}
\usage{
join_worms_spp(spp, check_cache = TRUE)
}
\arguments{
\item{spp}{A species data frame from \code{\link[=get_spp_names]{get_spp_names()}}}
}
\value{
A cleaned data frame
}
\description{
Join WoRMS data to species data frame
}
