% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{make_grid}
\alias{make_grid}
\title{Make prediction grid over years and survey grid}
\usage{
make_grid(survey_grid, years)
}
\arguments{
\item{survey_grid}{A dataframe from \code{\link[=prep_stitch_grids]{prep_stitch_grids()}}}

\item{years}{A numeric vector of years}
}
\value{
A dataframe with as many rows as \code{nrow(survey_grid) * length(years)}
}
\description{
Make prediction grid over years and survey grid
}
