% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-pages.R
\name{make_pages}
\alias{make_pages}
\title{Make synopsis pages}
\usage{
make_pages(
  dat,
  dat_iphc,
  spp,
  aspect = 1.35,
  short_page_height_ratio = 0.78,
  width = 11.5,
  debug = FALSE,
  resolution = 170,
  png_format = TRUE,
  spp_file = clean_name(spp),
  report_folder = "report",
  report_lang_folder = "report",
  tag = "main",
  include_map_square = FALSE,
  map_xlim = c(360, 653),
  map_ylim = c(5275, 6155),
  save_gg_objects = FALSE,
  survey_cols = c(`SYN WCHG` = "#E41A1C", `SYN WCVI` = "#984EA3", `SYN HS` = "#377EB8",
    `SYN QCS` = "#4DAF4A", `HBLL OUT N` = "#FF7F00", `HBLL OUT S` = "#FDBF6F",
    `HBLL INS N/S` = "#A65628", `IPHC FISS` = "#F781BF", `MSSM WCVI` = "#6c6c6c",
    `MSSM Geostat` = "#6c6c6c", `MSA HS` = "#6c6c6c", `SYN HS/QCS/WCHG/WCVI` = "#6c6c6c",
    `SYN HS/QCS/WCVI` = "#6c6c6c", `HBLL OUT N/S` = "#6c6c6c", Commercial = "#000000"),
  mat_min_n = 20,
  survey_map_outlier = 1,
  synoptic_max_survey_years = list(`SYN WCHG` = 2020, `SYN HS` = 2019, `SYN WCVI` = 2018,
    `SYN QCS` = 2019),
  hbll_out_max_survey_years = list(`HBLL OUT N` = 2019, `HBLL OUT S` = 2020),
  parallel = FALSE,
  french = FALSE,
  final_year_comm = 2021,
  final_year_surv = 2022,
  length_ticks = NULL,
  index_ggplot,
  spatiotemporal_cpue = FALSE,
  raw_cpue = NULL,
  shapefile = NULL,
  return_data = FALSE
)
}
\arguments{
\item{dat}{A data list object from \code{\link[gfdata:cache_pbs_data]{gfdata::cache_pbs_data()}}.}

\item{dat_iphc}{A data list object from \code{\link[gfiphc:cache_pbs_data_iphc]{gfiphc::cache_pbs_data_iphc()}}.}

\item{spp}{A species common name.}

\item{aspect}{The aspect ratio of the 2nd page.}

\item{short_page_height_ratio}{The aspect ratio of the shorter first page.}

\item{width}{The width of a page.}

\item{debug}{If \code{TRUE} then a debugging grid will be added to the plot
layout.}

\item{resolution}{The resolution for a png page.}

\item{png_format}{If \code{TRUE} then the pages will be png instead of PDF.}

\item{spp_file}{A version of the species name to be used when generating file
names.}

\item{report_folder}{The folder in which to create the report.}

\item{include_map_square}{If \code{TRUE} then a perfect square will be added to
the maps to help adjust the figure layout to get the aspect ratio of these
maps correct.}

\item{map_xlim}{The X axis limits in UTM coordinates.}

\item{map_ylim}{The Y axis limits in UTM coordinates.}

\item{save_gg_objects}{If \code{TRUE} then the ggplot2 objects will be saved as a
list object in \code{file.path(report_folder, 'ggplot-objects')}. This will also
cause the CPUE model to be cached.}

\item{survey_cols}{A vector of colors for the surveys.}

\item{mat_min_n}{The minimum number of samples before the maturity ogives are
plotted.}

\item{survey_map_outlier}{A quantile above which of the colors in the map
should be squashed the same color. Useful to avoid a small number of
outlying cells from distorting the color scale.}

\item{synoptic_max_survey_years}{A vector of length 2 giving the maximum
synoptic survey years to include.}

\item{parallel}{Parallel CPUE index fitting?}

\item{length_ticks}{A data frame indicating optional length composition
x-axis breaks and labels.}

\item{index_ggplot}{Pre-made ggplots for survey indices.}

\item{spatiotemporal_cpue}{Logical: use commercial trawl CPUE sdmTMB model output?}

\item{shapefile}{An optional shapefile used to subset or highlight a region
of interest.}

\item{return_data}{Return the data used to make the plots? Defaults to \code{FALSE}.}
}
\value{
This function generates 2 png files with all of the plots for a given species.
}
\description{
This is the main workhorse function that creates the synopsis pages.
}
