% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{plot_catches}
\alias{plot_catches}
\title{Plot catches}
\usage{
plot_catches(dat, blank_plot = FALSE, xlim = c(1955, 2017), ...)
}
\arguments{
\item{dat}{The data}

\item{blank_plot}{Whether or not the plot should be blank}

\item{xlim}{x limit}

\item{...}{Other arguments to pass to \code{\link[gfplot:plot_catch]{gfplot::plot_catch()}}.}
}
\description{
Plot catches
}
