% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-plots.R
\name{plot_cpue_spaghetti}
\alias{plot_cpue_spaghetti}
\title{Plot locality specific estimates from a CPUE standardization model}
\usage{
plot_cpue_spaghetti(model, fleet, index_data, era = c("modern",
  "historical"), the_formula_version = "Full standardization")
}
\arguments{
\item{model}{Output from \code{\link[gfplot:fit_cpue_index_glmmtmb]{gfplot::fit_cpue_index_glmmtmb()}}.}

\item{fleet}{The original data object. Output from
\code{\link[gfplot:tidy_cpue_historical]{gfplot::tidy_cpue_historical()}} or \code{\link[gfplot:tidy_cpue_index]{gfplot::tidy_cpue_index()}}.}

\item{index_data}{Standardized index data. Output from
\code{\link[gfplot:predict_cpue_index_tweedie]{gfplot::predict_cpue_index_tweedie()}}.}

\item{era}{Modern or historical.}

\item{the_formula_version}{A formula version to filter from \code{index_data} if
there is than one.}
}
\description{
Plot locality specific estimates from a CPUE standardization model
}
