% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pstar.R
\name{plot_pstar}
\alias{plot_pstar}
\title{Plot pstar - Visualization of pstar Estimation}
\usage{
plot_pstar(pstar_object)
}
\arguments{
\item{pstar_object}{A list containing the following components:
\itemize{
\item \code{pred_df}: A data frame with predicted values and confidence intervals.
\item \code{pstar_df}: A data frame with the estimated pstar value.
\item \code{species}: A character vector containing the species name.
}}
}
\value{
A ggplot2 object displaying the pstar visualization.
}
\description{
Generates a visualization of pstar estimation, including a line plot of
predicted values, shaded ribbons representing confidence intervals, and a
vertical line indicating the estimated pstar value.
}
