% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_survey_maps}
\alias{plot_survey_maps}
\title{Plot survey maps}
\usage{
plot_survey_maps(pred_dat, raw_dat, show_axes = FALSE,
  show_raw_data = TRUE, pos_pt_col = "#FFFFFF60",
  bin_pt_col = "#FFFFFF40", pos_pt_fill = "#FFFFFF05",
  north_symbol = FALSE, fill_column = "combined", trans = "sqrt",
  show_model_predictions = TRUE, annotations = c("SYN", "IPHC",
  "HBLL"), syn_qcs_hs_year = 2017, syn_wcvi_wchg_year = 2018,
  hbll_n_year = 2017, hbll_s_year = 2016, iphc_year = 2017, ...)
}
\arguments{
\item{pred_dat}{The \code{pred_dat} element from \code{\link[=fit_survey_maps]{fit_survey_maps()}}.}

\item{raw_dat}{The \code{raw_dat} element from \code{\link[=fit_survey_maps]{fit_survey_maps()}}.}

\item{show_axes}{Logical for whether or not to show axes.}

\item{show_raw_data}{Logical for whether or not show the raw data.}

\item{pos_pt_col}{The colour for the positive points.}

\item{bin_pt_col}{The colour for the occurrence/binary points.}

\item{pos_pt_fill}{The fill colour for the positive points.}

\item{north_symbol}{Logical for whether or not to include the North symbol.}

\item{fill_column}{A character object describing the fill column name.}

\item{trans}{The transformation function to apply to the color scale. Passed
to \code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}.}

\item{show_model_predictions}{Logical for whether or not to show the model
predictions.}

\item{annotations}{A character object describing which annotations to
include. These annotations include the relevent years on the map.}

\item{syn_qcs_hs_year}{Year for a synoptic survey label.}

\item{syn_wcvi_wchg_year}{Year for a synoptic survey label.}

\item{hbll_n_year}{Year for a HBLL N survey label.}

\item{hbll_s_year}{Year for a HBLL S survey label.}

\item{iphc_year}{Year for IPHC survey label.}

\item{...}{Any other arguments to pass to \code{\link[gfplot:plot_survey_sets]{gfplot::plot_survey_sets()}}.}
}
\description{
\code{plot_survey_maps()} is a helper function that calls
\code{\link[gfplot:plot_survey_sets]{gfplot::plot_survey_sets()}} appropriately for the various surveys.
}
