% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-iphc.R
\name{prep_iphc_stitch_dat}
\alias{prep_iphc_stitch_dat}
\title{Prepare IPHC FISS data for index stitching}
\usage{
prep_iphc_stitch_dat()
}
\value{
A dataframe the same length as \code{survey_dat}
}
\description{
Prepare IPHC FISS data for index stitching
}
