% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{prep_mssm_dat}
\alias{prep_mssm_dat}
\title{Prepare Multispecies Small Mesh (MSSM) survey set data for index stitching}
\usage{
prep_mssm_dat(survey_dat)
}
\arguments{
\item{survey_dat}{A dataframe from \verb{[gfplot::get_survey_sets()]}}
}
\value{
A dataframe the same length as \code{survey_dat}
}
\description{
Prepare Multispecies Small Mesh (MSSM) survey set data for index stitching
}
