% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{prep_stitch_dat}
\alias{prep_stitch_dat}
\title{Prepare survey set data for index stitching}
\usage{
prep_stitch_dat(survey_dat, bait_counts)
}
\arguments{
\item{survey_dat}{A dataframe from \verb{[gfplot::get_survey_sets()]}}

\item{bait_counts}{A dataframe from \code{\link[=get_ll_bait_counts]{get_ll_bait_counts()}}}
}
\value{
A dataframe the same length as \code{survey_dat}
}
\description{
Prepare survey set data for index stitching
}
