% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{prep_stitch_grids}
\alias{prep_stitch_grids}
\title{Write grids used for stitching index}
\usage{
prep_stitch_grids(
  grid_dir = file.path("data-outputs", "grids"),
  hbll_ins_grid_input = file.path("data-cache", "hbll-inside-grid.rds")
)
}
\description{
Write grid objects used for stitching index. These are written to a
<data-outputs/grids> directory.
Note: Might not be needed if clean grids are added to \code{gfdata}
}
