% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rougheye_split}
\alias{rougheye_split}
\title{Split rougheye species name}
\usage{
rougheye_split(x)
}
\arguments{
\item{x}{Latin species name}
}
\description{
Split rougheye species name
}
\examples{
rougheye_split("sebastes aleutianus/melanostictus complex")
}
