% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{scale_indices}
\alias{scale_indices}
\title{Scale Indices}
\usage{
scale_indices(
  dat_design = NULL,
  dat_geostat = NULL,
  .groups = c("species_common_name", "survey_abbrev"),
  scale_by = c("upperci", "geomean")
)
}
\arguments{
\item{dat_design}{A data frame containing the design-based indices.}

\item{dat_geostat}{A data frame containing the output from
\code{sdmTMB::get_index()}.}

\item{.groups}{A character vector specifying the grouping variables to use
for scaling. Default is \code{c("species_common_name", "survey_abbrev")}.}

\item{scale_by}{A character string specifying the method to scale by, with
the following cases:
\itemize{
\item If both \code{dat_design} and \code{dat_geostat} are provided:
\itemize{
\item \code{"upperci"}: Indices will be scaled by the maximum scaled
upper confidence interval (\code{upperci}), which itself is scaled
by the geometric mean of the geostatistical index.
\item \code{"geomean"}: Indices will be scaled by the scaled geometric
mean of the geostatistical index.
}
\item If only one of \code{dat_design} or \code{dat_geostat} is provided:
\itemize{
\item \code{"upperci"}:w Indices will be scaled by the maximum
\code{upperci} of the available index.
\item \code{"geomean"}: Indices will be scaled by the geometric mean
of the available index.
}
}}
}
\value{
A data frame similar to \code{dat_design}, with additional scaled
columns such as \code{biomass_scaled}, \code{lowerci_scaled}, and
\code{upperci_scaled}.
}
\description{
This function scales index values from design-based, geostatistical model-
based, or the combination of both for plotting.
}
\details{
This function scales index values based on specified methods, grouping by
specified columns. If no valid data is available (e.g., all \code{NA} values
in the relevant column), scaled values will be set to \code{NA_real_}.
}
\examples{
# Example?
}
