% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-filtering.R
\name{subset_spatial}
\alias{subset_spatial}
\title{Subset a data frame spatially using an \pkg{sf} polygon}
\usage{
subset_spatial(dat, sf_poly, xy_coords, dat_crs = 4326, return_sf = FALSE)
}
\arguments{
\item{dat}{A data frame (not \pkg{sf} object) containing spatial data with
longitude and latitude coordinates.}

\item{sf_poly}{An \pkg{sf} polygon object used to spatially filter the dataset.}

\item{xy_coords}{A character vector of length two specifying the
column names for longitude and latitude in \code{dat}.}

\item{dat_crs}{Integer. Coordinate reference system (CRS) for \code{dat}.
Default is \code{4326} (WGS84).}

\item{return_sf}{Logical. If \code{TRUE}, returns an \pkg{sf} object; if
\code{FALSE}, returns a data frame. Default is \code{FALSE}.}
}
\value{
A dataframe without geometry (default) or an \pkg{sf} object if
\code{return_sf = TRUE}.
}
\description{
Filters a dataset based on spatial intersection with a given polygon.
The function converts the input data frame to an \pkg{sf} object using
specified coordinate columns, applies coordinate transformations if
necessary, and returns either a filtered dataframe (default) or an
\pkg{sf} object.
}
