% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostatistical-index.R
\name{fit_sdmTMB_westcoast}
\alias{fit_sdmTMB_westcoast}
\title{Fit a geostatistical index standardization model to survey data}
\usage{
fit_sdmTMB_westcoast(
  species_rds,
  survey,
  species_name = "",
  n_knots = 200,
  cell_width = 2,
  anisotropy = FALSE,
  silent = TRUE,
  bias_correct = FALSE,
  include_depth = FALSE
)
}
\arguments{
\item{species_rds}{File path to an .rds file from \code{\link[gfplot:get_data]{gfplot::cache_pbs_data()}}.}

\item{survey}{Survey abbreviation for the survey to fit.}

\item{species_name}{Species name. Only used in the output to help with processing.}

\item{n_knots}{Number of knots in the SPDE approximation.}

\item{cell_width}{Sell width for the prediction grid.}

\item{anisotropy}{Logical for anisotropy.}

\item{silent}{Logical for verbosity of output.}

\item{bias_correct}{Logical for bias correction from TMB.}

\item{include_depth}{Logical for whether to include depth in the model.}
}
\value{
A list object.
}
\description{
Fit a geostatistical index standardization model to survey data
}
