% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Get all data}
\usage{
get_data(type = c("A", "B"), path = ".", compress = FALSE,
  force = FALSE)
}
\arguments{
\item{type}{A or B}

\item{path}{Path}

\item{compress}{TRUE or FALSE to compress .rds}

\item{force}{Should data be downloaded even if already cached data exists?}
}
\description{
Get all data
}
