% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-iphc.R
\name{get_iphc_stitched_index}
\alias{get_iphc_stitched_index}
\title{Get index using \code{sdmTMB} for IPHC surveys}
\usage{
get_iphc_stitched_index(
  survey_dat,
  species,
  model_type = "st-rw",
  form = NULL,
  family = sdmTMB::tweedie(),
  time = "year",
  spatial = "on",
  spatiotemporal = "rw",
  time_varying = NULL,
  time_varying_type = NULL,
  mesh = NULL,
  cutoff = 20,
  offset = "log_eff_skate",
  priors = sdmTMB::sdmTMBpriors(),
  silent = TRUE,
  ctrl = sdmTMB::sdmTMBcontrol(),
  gradient_thresh = 0.001,
  cache = NULL,
  check_cache = FALSE,
  cache_fits = FALSE,
  cache_predictions = FALSE,
  grid
)
}
\arguments{
\item{survey_dat}{A data frame from \code{\link[=prep_iphc_stitch_dat]{prep_iphc_stitch_dat()}}.}

\item{species}{A string specifying the \code{species_common_name}.}

\item{model_type}{An suffix to the filename indicating model type (default = "st-rw").}

\item{form}{Optional string specifying model formula.
'catch ~ 1' (the default, unless \code{family = poisson()} or \code{family = sdmTMB::censored_possion()}
then "catch ~ 1 + (1|obs_id)").}

\item{family}{The family and link for \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{time}{An optional time column name (as character), used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
(default = 'year')}

\item{spatial}{Estimate spatial random fields? See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
Default is 'rw' if `model_type = "st-rw" or "st-rw_tv-rw".}

\item{spatiotemporal}{Estimate the spatiotemporal random fields, see \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
Default is 'rw'   if `model_type = "st-rw" or "st-rw_tv-rw".}

\item{time_varying}{An optional one-sided formula describing covariates that
should be modelled as a time-varying process. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.
Default is \code{NULL}.}

\item{time_varying_type}{Type of time-varying process to apply to
‘time_varying’ formula. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}. Default is \code{NULL}.}

\item{mesh}{Optional mesh object created using \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{cutoff}{If \code{mesh = NULL}, mesh cutoff for \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{offset}{A string naming the offset column in \code{dat} used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}
(default = 'log_eff_skate').}

\item{priors}{Optional penalties/priors used in \code{\link[sdmTMB:priors]{sdmTMB::sdmTMBpriors()}}.}

\item{silent}{A boolean. Silent or include optimization details.}

\item{ctrl}{Optimization control options via \code{\link[sdmTMB:sdmTMBcontrol]{sdmTMB::sdmTMBcontrol()}}.}

\item{gradient_thresh}{Threshold used in \code{\link[sdmTMB:sanity]{sdmTMB::sanity()}} (default = 0.001).}

\item{cache}{A string specifying file path to cache directory.}

\item{check_cache}{Check whether index file already exists? Default = \code{FALSE}.}

\item{cache_fits}{Cache model fits? Can be large.}

\item{cache_predictions}{Cache model predictions? Can be large.}

\item{grid}{A data frame containing the locations of IPHC stations used to make predictions and generate the index.}
}
\value{
Either a string or dataframe:
\itemize{
\item \verb{insufficient data to stitch regions} if the number of positive sets is too low to stitch
\item \verb{Failed sanity check} if the model failed to converge
\item A dataframe containing the stitched index formatted to use with \code{\link[gfplot:plot_survey_index]{gfplot::plot_survey_index()}}
}
}
\description{
Get index using \code{sdmTMB} for IPHC surveys
}
