% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{get_stitched_index}
\alias{get_stitched_index}
\title{Get stitched index across survey regions in synoptic trawl and HBLL surveys}
\usage{
get_stitched_index(
  survey_dat,
  species = "arrowtooth flounder",
  survey_type = "synoptic",
  model_type = "st-rw",
  mesh = NULL,
  cutoff = 20,
  family = sdmTMB::tweedie(),
  offset = "offset",
  silent = TRUE,
  ctrl = sdmTMB::sdmTMBcontrol(nlminb_loops = 1L, newton_loops = 1L),
  cache = file.path("report", "stitch-cache")
)
}
\arguments{
\item{survey_dat}{A dataframe from \code{\link[=prep_stitch_dat]{prep_stitch_dat()}}.}

\item{species}{A string specifying the \code{species_common_name}.}

\item{survey_type}{A string matching one of: "synoptic" (the default), "hbll_outside", "hbll_inside".}

\item{model_type}{A string matching one of: "st-rw" (the default), "st-rw_tv-rw".}

\item{mesh}{Optional mesh object created using \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{cutoff}{If \code{mesh = NULL}, mesh cutoff for \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{family}{The family and link for \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{offset}{A string naming the offset column in \code{dat} used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{silent}{A boolean. Silent or include optimization details.}

\item{ctrl}{Optimization control options via \code{\link[sdmTMB:sdmTMBcontrol]{sdmTMB::sdmTMBcontrol()}}.}

\item{cache}{A string specifying file path to cache directory.}
}
\value{
Either a string or dataframe:
\itemize{
\item \verb{insufficient data to stitch regions} if the number of positive sets is too low to stitch
\item \verb{Failed sanity check} if the model failed to converge
\item A dataframe containing the stitched index formatted to use with \code{\link[gfplot:plot_survey_index]{gfplot::plot_survey_index()}}
}
}
\description{
Get stitched index across survey regions in synoptic trawl and HBLL surveys
}
