% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{get_stitched_index}
\alias{get_stitched_index}
\title{Get stitched index across survey regions in synoptic trawl and HBLL surveys}
\usage{
get_stitched_index(
  survey_dat,
  species = "arrowtooth flounder",
  survey_type = "synoptic",
  model_type = "st-rw",
  form = NULL,
  family = sdmTMB::tweedie(),
  time = "year",
  spatial = "on",
  spatiotemporal = "rw",
  time_varying = NULL,
  time_varying_type = NULL,
  mesh = NULL,
  cutoff = 20,
  offset = "offset",
  extra_time = NULL,
  priors = sdmTMB::sdmTMBpriors(),
  silent = TRUE,
  ctrl = sdmTMB::sdmTMBcontrol(nlminb_loops = 1L, newton_loops = 1L),
  gradient_thresh = 0.001,
  upr = NULL,
  cache = NULL,
  check_cache = FALSE,
  survey_grid = NULL,
  grid_dir
)
}
\arguments{
\item{survey_dat}{A data frame from \code{\link[=prep_stitch_dat]{prep_stitch_dat()}}.}

\item{species}{A string specifying the \code{species_common_name}.}

\item{survey_type}{A string matching one of: "synoptic" (the default), "hbll_outside", "hbll_inside".}

\item{model_type}{A string matching one of: "st-rw" (the default), "st-rw_tv-rw".}

\item{family}{The family and link for \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}.}

\item{mesh}{Optional mesh object created using \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{cutoff}{If \code{mesh = NULL}, mesh cutoff for \code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}.}

\item{offset}{A string naming the offset column in \code{dat} used in \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{silent}{A boolean. Silent or include optimization details.}

\item{ctrl}{Optimization control options via \code{\link[sdmTMB:sdmTMBcontrol]{sdmTMB::sdmTMBcontrol()}}.}

\item{gradient_thresh}{Threshold used in \code{\link[sdmTMB:sanity]{sdmTMB::sanity()}} (default = 0.001).}

\item{cache}{A string specifying file path to cache directory.}

\item{survey_grid}{A data frame containing the spatial grid over which predictions are to be made.
If \code{survey_grid} = NULL (the default). Grid should contain cell area.}

\item{grid_dir}{Path where cleaned grids were stored from \code{\link[=prep_stitch_grids]{prep_stitch_grids()}}}
}
\value{
Either a string or dataframe:
\itemize{
\item \verb{insufficient data to stitch regions} if the number of positive sets is too low to stitch
\item \verb{Failed sanity check} if the model failed to converge
\item A dataframe containing the stitched index formatted to use with \code{\link[gfplot:plot_survey_index]{gfplot::plot_survey_index()}}
}
}
\description{
Get stitched index across survey regions in synoptic trawl and HBLL surveys
}
