% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfsynopsis.R
\docType{package}
\name{gfsynopsis}
\alias{gfsynopsis-package}
\alias{gfsynopsis}
\title{gfsynopsis package}
\description{
gfsynopsis package
}
\details{
See the README on
\href{https://github.com/seananderson/gfsynopsis#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seananderson/gfsynopsis}
  \item Report bugs at \url{https://github.com/seananderson/gfsynopsis/issues}
}

}
\author{
\strong{Maintainer}: Sean Anderson \email{sean.anderson@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Elise Keppel
}

}
