% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_species_metadata.R
\name{join_itis_spp}
\alias{join_itis_spp}
\title{Join ITIS data to species data frame}
\usage{
join_itis_spp(spp)
}
\arguments{
\item{spp}{A species data frame from \code{\link[=get_spp_names]{get_spp_names()}}}
}
\value{
A cleaned data frame
}
\description{
Join ITIS data to species data frame
}
\examples{
get_spp_names() |>
  join_itis_spp()
}
