% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-pages.R
\name{make_pages}
\alias{make_pages}
\title{Make pages}
\usage{
make_pages(dat, spp, aspect = 1.35, short_page_height_ratio = 0.78,
  width = 11.5, debug = FALSE, resolution = 220, png_format = TRUE,
  spp_file = gfsynopsis:::clean_name(spp), report_folder = "report",
  include_map_square = FALSE, map_xlim = c(360, 653), map_ylim = c(5275,
  6155), survey_cols = c(RColorBrewer::brewer.pal(7L, "Set2"), "#303030",
  "#a8a8a8", "#a8a8a8", "#a8a8a8"), survey_col_names = c("SYN WCHG", "SYN HS",
  "SYN QCS", "SYN WCVI", "HBLL OUT N", "HBLL OUT S", "IPHC FISS", "Commercial",
  "HBLL INS N", "HBLL INS S", "MSA HS"))
}
\arguments{
\item{dat}{TODO}

\item{spp}{TODO}

\item{aspect}{TODO}

\item{width}{TODO}

\item{debug}{TODO}

\item{resolution}{TODO}

\item{png_format}{TODO}

\item{spp_file}{TODO}

\item{report_folder}{TODO}

\item{survey_cols}{TODO}

\item{survey_col_names}{TODO}
}
\description{
Make pages
}
