% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange.R, R/make-pages.R
\name{make_pages}
\alias{make_pages}
\title{Make pages}
\usage{
make_pages(dat, spp, output_path, aspect = 1.35, width = 11.5,
  survey_cols = gg_color_hue(7L), debug = FALSE,
  ageing_method_codes = c(3, 17), survey_age_year_range = c(2002,
  as.numeric(format(Sys.Date(), "\%Y"))))

make_pages(dat, spp, output_path, aspect = 1.35, width = 11.5,
  survey_cols = gg_color_hue(7L), debug = FALSE,
  ageing_method_codes = c(3, 17), survey_age_year_range = c(2002,
  as.numeric(format(Sys.Date(), "\%Y"))))
}
\arguments{
\item{dat}{TODO}

\item{spp}{TODO}

\item{output_path}{TODO}

\item{aspect}{TODO}

\item{width}{TODO}

\item{survey_cols}{TODO}

\item{debug}{TODO}

\item{dat}{TODO}

\item{spp}{TODO}

\item{output_path}{TODO}

\item{aspect}{TODO}

\item{width}{TODO}

\item{survey_cols}{TODO}

\item{debug}{TODO

importFrom grDevices dev.off pdf}
}
\description{
Make pages

Make pages
}
\examples{
\dontrun{
library("gfplot")
dc <- "data-cache"
dir.create(dc, showWarnings = FALSE)
dir.create("figs", showWarnings = FALSE)
cache_pbs_data(spp, path = dc)

dat <- list()
dat$survey_tows     <- readRDS(file.path(dc, "pbs-surv-tows.rds"))
dat$survey_samples  <- readRDS(file.path(dc, "pbs-surv-samples.rds"))
dat$comm_samples    <- readRDS(file.path(dc, "pbs-comm-samples.rds"))
dat$catch           <- readRDS(file.path(dc, "pbs-catch.rds"))
dat$cpue_spatial    <- readRDS(file.path(dc, "pbs-cpue-spatial.rds"))
dat$cpue_spatial_ll <- readRDS(file.path(dc, "pbs-cpue-spatial-ll.rds"))
dat$survey_index    <- readRDS(file.path(dc, "pbs-surv-index.rds"))
dat$age_precision   <- readRDS(file.path(dc, "pbs-age-precision.rds"))

make_pages(dat, "arrowtooth flounder", output_path = "figs")
}
\dontrun{
library("gfplot")
spp <- "arrowtooth flounder"
dc <- "data-cache"
dir.create(dc, showWarnings = FALSE)
dir.create("figs", showWarnings = FALSE)
cache_pbs_data(spp, path = dc)

dat <- list()
dat$survey_tows     <- readRDS(file.path(dc, "pbs-surv-tows.rds"))
dat$survey_samples  <- readRDS(file.path(dc, "pbs-surv-samples.rds"))
dat$comm_samples    <- readRDS(file.path(dc, "pbs-comm-samples.rds"))
dat$catch           <- readRDS(file.path(dc, "pbs-catch.rds"))
dat$cpue_spatial    <- readRDS(file.path(dc, "pbs-cpue-spatial.rds"))
dat$cpue_spatial_ll <- readRDS(file.path(dc, "pbs-cpue-spatial-ll.rds"))
dat$survey_index    <- readRDS(file.path(dc, "pbs-surv-index.rds"))
dat$age_precision   <- readRDS(file.path(dc, "pbs-age-precision.rds"))

make_pages(dat, spp, output_path = "figs")
}
}
