% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{plot_cpue_indices}
\alias{plot_cpue_indices}
\title{Plot CPUE indices}
\usage{
plot_cpue_indices(
  dat,
  blank_plot = FALSE,
  xlim = c(1996, 2017),
  area_labels = c("3CD5ABCDE", "5CDE", "5AB", "3CD")
)
}
\arguments{
\item{dat}{The data}

\item{blank_plot}{Whether or not a blank plot should be generated}

\item{xlim}{The x limits}

\item{area_labels}{Area labels}
}
\description{
Plot CPUE indices
}
