% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch-survey.R
\name{prep_stitch_dat}
\alias{prep_stitch_dat}
\title{Prepare survey set data for index stitching}
\usage{
prep_stitch_dat(spp_dat, bait_count_path)
}
\arguments{
\item{spp_dat}{A dataframe from \code{\link[gfplot:get_survey_sets]{gfplot::get_survey_sets()}}}
}
\value{
A dataframe the same length as \code{spp_dat}
}
\description{
Prepare survey set data for index stitching
}
